pyproject.toml
[project]
name = "excel-to-ocr"
version = "1.0.0"
description = "Convert complex Excel files with many columns to OCR-friendly formats (Markdown/PDF)"
authors = [
    {name = "TanishaR2", email = "tanisha.ramani@simform.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["excel", "ocr", "pdf", "markdown", "data-extraction", "financial-models"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]

dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "pypdf2>=3.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/TanishaR2/extracting-from-excel"
Repository = "https://github.com/TanishaR2/extracting-from-excel"
Issues = "https://github.com/TanishaR2/extracting-from-excel/issues"
Documentation = "https://github.com/TanishaR2/extracting-from-excel/blob/main/docs/README.md"

[project.scripts]
excel-to-ocr = "run_enhanced:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/excel_to_ocr"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
